requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx[socks]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0
playwright>=1.40.0
